Revert to master version of AkCollatorFactory
There were still problems, and this is just an attempt to make the
wording of an explain being consistent, so rather than going down
the rabbit hole trying to fix the bugs, I'm just going to change
the test so that it resets. Chances are a user won't be using a
variety of collate names that amount to the same collation, and
then need consistent output in an explain.