Address three found bugs with the lookahead processing not managing the cursor state properly.
Intersect_Ordered#nextBindings() needed to be smarter about handling the
underlying bindings as the lookahead may have moved the input binding
forward by several leaps depending upon lookahead. 

Intersect_Ordered#cancelBindings() needed to mark the cursor as closed
to force the parent operator to re-open it, so it would correctly reset
the two input operators. 

Map_NestedLoops - Two places to check if the input cursor is open
already, and if not make sure it is before reading from it. 

Also, update some of the debug messages to make it clearer where (and
why) the debug messages are appearing. Do the same for several asserts
as well.