The great Schema purge.
As the code moves toward relying more on RowType, remove a number of
places caching the row type schema, especially in the StoreAdpater. 

In some cases this schema was used as a shortcut to get the AIS, for
which I've added the storeAdapter.getAIS() to resolve. In other it's
easier to use SchemaCache.getGlobalSchema() which in turn retrives it
from the AIS where it is kept. 

Unit tests pass, but no attempt to run ITs yet.