Look for a compatible index before creating one for FORIEN KEYs.
Do this by using the same method that was already there to find
referenced side indexes but add a requireUnique parameter.

This avoids creating an index if a compatible one already exists, where
compatible means starting with the required columns in some order.

Includes some cleanup when printing out constraints for info schem as
there is no longer a 1-1 mapping between index and FK.

As there is no way to no, an index that was created implicitly is no
longer dropped if the FK is dropped.